Rename ProjectEnvironment to LegacyProjectEnvironment
